modularization of the code base cont.
