+++ source/engine/main-dedicated.mjs
@@ -8,13 +8,19 @@ import Con from './common/Console.mjs';
 import Host from './common/Host.mjs';
 import V from './client/V.mjs';
 import NET from './network/Network.mjs';
-import SV from './network/Server.mjs';
+import SV from './server/Server.mjs';
+import PR from './server/Progs.mjs';
+import Mod from './common/Mod.mjs';
+import * as WebSocket from 'ws';
 
 export default class EngineLauncher {
   static async Launch() {
     console.log('Launching engine as dedicated server...');
 
     registry.isDedicatedServer = true;
+
+    // inject some external dependencies
+    registry.WebSocket = WebSocket;
 
     // hooking up all required components
     registry.Sys = Sys;
@@ -25,6 +31,8 @@ export default class EngineLauncher {
     registry.V = V;
     registry.NET = NET;
     registry.SV = SV;
+    registry.PR = PR;
+    registry.Mod = Mod;
 
     try {
       await Sys.Init();
