+++ source/game/id1/entity/Subs.mjs
@@ -1,4 +1,4 @@
-/* global Vector */
+import Vector from '../../../shared/Vector.mjs';
 
 import { channel, moveType, solid } from "../Defs.mjs";
 import { EntityWrapper } from "../helper/MiscHelpers.mjs";
