+++ source/game/id1/main.mjs
@@ -1,6 +1,3 @@
-// eslint-disable-next-line no-global-assign
-try { self = Object.freeze({}); } catch (e) { console.debug(e); } // CR: self is poison
-
 import { ServerGameAPI } from './GameAPI.mjs';
 import { ClientGameAPI } from './client/ClientAPI.mjs';
